safari-600.7-branch/LayoutTests/inspector-obsolete/http-tests/network/network-size-chunked.html

<html>
<head>
<script src="../inspector-test.js"></script>
<script>
function loadImages()
{
    var image = document.createElement("img");
    image.onload = resourceLoaded;
    image.setAttribute("src", "resources/resource.php?type=image&size=500&gzip=true&chunked=true");
    document.body.appendChild(image);
}

function resourceLoaded()
{
    console.log("Done.");
}

function test()
{
    InspectorTest.addConsoleSniffer(step2);
    InspectorTest.evaluateInPage("loadImages()");

    function step2()
    {
        var request1 = WebInspector.panel("network").requests[WebInspector.panel("network").requests.length - 1];
        InspectorTest.addResult(request1.url);
        InspectorTest.addResult("resourceSize: " + request1.resourceSize);
        InspectorTest.addResult("transferSize: " + request1.transferSize);

        InspectorTest.completeTest();
    }
}
</script>
</head>
<body onload="runTest()">
<p>Tests network resource size reporting for chunked transfer encoding.</p>
</body>
</html>
